uses modules to complete plugin core behavior
calling M.toggle() while currently in a ruby buffer will not switch
users between spec and app files. It will create/generate files if they
are missing.

It does this in a somewhat clunky way. For example, if you are on a
request spec and toggle, you will land on a new empty file in the
app/request directory. This is probably not what you want.